fix: overwrite stale recording outputs
